requirements.txt
gql[requests]
python-dotenv
